+++ src/js/modules/_editor.js
@@ -1,9 +1,39 @@
 export default Editor;
 import Sortable from '../lib/_Sortable';
 import DomEl from '../lib/_DomEl';
+import DomButton from '../lib/_DomButton';
 
 let Editor = function(containerEl, data) {
     let Blocks = [];
+    let BlockChooser = {
+        choiceDiv: new DomEl('div.blockChoices'),
+        create: function() {
+            let blockChoices = window.Hat.getBlocks();
+            for (let [slug, block] of Object.entries(blockChoices)) {
+                let button = new DomButton(block.description, block.icon, 'choiceBtn', block.name); 
+                button.dataset.slug = slug;
+                BlockChooser.choiceDiv.append(button);
+                let control = Interface;
+                button.addEventListener('click', function(e) {
+                    BlockChooser.toggle();
+                    let button = e.target.closest('button');
+                    control.addBlock(true, false, button.dataset.slug);
+                });
+            };
+            Elements.container.append(BlockChooser.choiceDiv);
+        },
+        toggle: function(originButton) {
+            let ChoiceDiv = BlockChooser.choiceDiv;
+            if (ChoiceDiv.classList.contains('show')) {
+                ChoiceDiv.classList.remove('show');
+            } else {
+                ChoiceDiv.style.left = originButton.offsetLeft + (originButton.offsetWidth/2) + 'px';
+                ChoiceDiv.style.top = originButton.offsetTop + originButton.offsetHeight + 'px';
+                ChoiceDiv.classList.add('show');
+                ChoiceDiv.children[0].focus();
+            }
+        }
+    };
     let BlockCount = 0;
     let Elements = {
         blockHolder: false,
@@ -17,12 +47,11 @@ let Editor = function(containerEl, data) {
     let Internal = {
         blockCount: 0,
         insertAddBlockButton: function() {
-            var a = new DomEl('a.block[title="Add a new block (and select it)][href="javascript:void(0)"]');
+            var a = new DomEl('a.block[title=Create a new block by clicking here. You will be taken to the block type selector][href="javascript:void(0)"]');
             var icon = new DomEl('i.fas.fa-plus-square');
             a.innerHTML = icon.outerHTML + ' Add block';
             a.addEventListener('click', function() {
-                //TODO: Show block list
-                Interface.addBlock();
+                BlockChooser.toggle(this);
             });
             Elements.container.append(a);
         },
@@ -39,6 +68,7 @@ let Editor = function(containerEl, data) {
                 Interface.addBlock();
             }
             Internal.manageSorting();
+            BlockChooser.create();
             document.execCommand('defaultParagraphSeparator', false, 'p');
         },
         manageSorting: function() {
@@ -55,7 +85,7 @@ let Editor = function(containerEl, data) {
     let Interface = {
         addBlock: function(focus=true,position=false, type=window.Hat.getDefault(), data=false) {
             let blockClass = window.Hat.getBlock(type);
-            let block = new blockClass(this, data);
+            let block = new blockClass.class(this, data);
             block.el.id = 'block' + new Date().getTime();
             if (position === false) {
                 Elements.blockHolder.appendChild(block.el);